Modules/AI/agents/MCP/JarvisMCP/requirements.txt
# JarvisMCP dependencies
fastmcp>=3.0.0
httpx>=0.25.0  # For HTTP client testing
python-dotenv>=1.0.0  # For .env configuration
